refactor: centralize tauri detection
