+++ src/lib/runtime.ts
@@ -1,12 +1,12 @@
 // src/lib/runtime.ts
-export const isTauri =
-  typeof window !== "undefined" &&
-  typeof (window as any).__TAURI_IPC__ === "function";
+import { isTauri } from "@/lib/db/sql";
+
+export { isTauri };
 
 export function requireTauri(message?: string) {
   if (!isTauri) {
     throw new Error(
-      message || "Tauri required: run via `npx tauri dev` to access SQLite"
+      message || "Tauri required: open the native window"
     );
   }
 }
