--- src/tauri/isTauri.ts
@@ -1,3 +1,0 @@
-export function isTauri(): boolean {
-  return typeof (window as any).__TAURI__ !== "undefined";
-}
