--- src/tauriEnv.ts
@@ -1,1 +1,0 @@
-export const isTauri = !!import.meta.env.TAURI_PLATFORM;
